Fix issue with SublimeText3
This commit simply comments out a print statement that I assume was used for debugging. It causes a plugin runtime error resulting in a non-functional enter key following a closing curly bracket.

For example, using this syntax as a test place the cursor right after the closing curly bracket and the enter key will appear to be unresponsive. 

export
    {
    const test = 20.0 &redef;
    }